# documentation/swagger.yaml
openapi: 3.0.0
info:
  title: MamaCheck API - Complete Reference
  description: |
    # MamaCheck - SMS-First Antenatal Care System

    Comprehensive REST API for pregnancy management, ANC reminders, and danger sign triage in Nigeria.

    ## Key Features
    - **Pregnancy Management**: Registration with OTP verification, timeline tracking
    - **ANC Reminders**: Daily scheduling at 07:00 WAT, personalized messages
    - **Danger Sign Triage**: Weekly check-ins with RED/YELLOW/GREEN outcomes
    - **CHEW Dashboard**: Real-time KPIs, women registry, red flag management
    - **Multi-language**: English, Pidgin, Yoruba, Hausa, Igbo
    - **SMS Integration**: Termii SMS gateway with DND compliance
    - **Visit Tracking**: Attendance marking with 10-minute undo window
    - **Reference Data**: LGA and PHC management with geolocation

    ## Security
    - JWT authentication for all protected endpoints
    - Role-based access (PATIENT, CHEW, SUPERVISOR, ADMIN)
    - OTP verification for phone number validation
    - Rate limiting on SMS and registration endpoints
    - BCRYPT password hashing in production

    ## SMS Triage Guide
    Women reply with symptom numbers:
    - **0** = No symptoms (GREEN)
    - **1** = Heavy bleeding (RED)
    - **2** = Severe headache (RED)
    - **3** = Swollen face/hands (RED)
    - **4** = Blurry vision (YELLOW)
    - **5** = Fever (YELLOW)
    - **6** = Reduced baby movement (YELLOW)
...
  version: "1.0.0"
...
servers:
...
paths:
...
  /dashboard/chew/overview:
    get:
      tags:
        - Dashboard
      summary: Get Dashboard Overview
      description: Real-time KPIs for CHEW dashboard
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Dashboard overview with KPIs
          content:
            application/json:
              schema:
                type: object
                properties:
                  summary:
                    type: object
                    properties:
                      totalWomen:
                        type: integer
                      activePregnancies:
                        type: integer
                      highRiskWomen:
                        type: integer
                      dueThisWeek:
                        type: integer
                      overdueVisits:
                        type: integer
                      redFlagsToday:
                        type: integer
                      redFlagsThisWeek:
                        type: integer
                      averageResponseTime:
                        type: number
                        description: Minutes
  /dashboard/chew/women:
    get:
      tags:
        - Dashboard
      summary: Get Women Registry
      description: List of assigned pregnant women with status
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Women registry
          content:
            application/json:
              schema:
                type: object
                properties:
                  total:
                    type: integer
                  page:
                    type: integer
                  women:
                    type: array
                    items:
                      $ref: "#/components/schemas/WomanSummary"
  /dashboard/chew/red-flags:
    get:
      tags:
        - Dashboard
      summary: Get Red Flags
      description: List of RED (critical) triage outcomes requiring CHEW follow-up
      security:
        - BearerAuth: []
      responses:
        "200":
          description: RED flag cases
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DangerReport"
  /dashboard/chew/weekly-summary:
    get:
      tags:
        - Dashboard
      summary: Get Weekly Summary
      description: Single-number overview for CHEW reporting
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Weekly statistics
          content:
            application/json:
              schema:
                type: object
                properties:
                  totalWomen:
                    type: integer
                  dueThisWeek:
                    type: integer
                  missedVisits:
                    type: integer
                  openRedFlags:
                    type: integer
                  weekStartDate:
                    type: string
                    format: date
                  weekEndDate:
                    type: string
                    format: date
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DangerReport:
      type: object
      properties:
        _id:
          type: string
        reportedSymptoms:
          type: array
          items:
            type: integer
            enum: [0, 1, 2, 3, 4, 5, 6, 7, 8]
        triageOutcome:
          type: string
          enum: [GREEN, YELLOW, RED]
        timestamp:
          type: string
          format: date-time
        requiresFollowup:
          type: boolean
        followup:
          type: object
          properties:
            status:
              type: string
              enum: [pending, in_progress, completed, escalated]
            outcome:
              type: string
              enum: [phone_contact, clinic_visit, referral_hospital]
    WomanSummary:
      type: object
      properties:
        _id:
          type: string
        name:
          type: string
        gestationalWeek:
          type: integer
        nextANCDue:
          type: string
          format: date
        redFlagStatus:
          type: boolean
...